feat(pre-root): safety-checks for pacman-extra

+++ setup/pre/root.sh
@@ -397,11 +397,13 @@ pacman_extra() {
         #   https://www.blackarch.org/downloads.html#install-repo
         local _blackarch="strap.sh"
         curl -O "https://blackarch.org/${_blackarch}"
-        chmod +x "${_blackarch}"
-        ./"${_blackarch}"
-        rm "${_blackarch}"
+        if ! (chmod +x "${_blackarch}" && ./"${_blackarch}"); then
+            printf "blackarch> failed, exiting"
+            exit 3
+        fi
 
         __separator
+        rm "${_blackarch}"
         __confirm "blackarch"
     }
 
@@ -411,9 +413,14 @@ pacman_extra() {
         # REF:
         #   https://wiki.archlinux.org/title/Unofficial_user_repositories#archzfs
         local _archzfs_key="DDF7DB817396A49B2A2723F7403BD972F75D9D76"
-        pacman-key --recv-keys "${_archzfs_key}"
-        pacman-key --finger "${_archzfs_key}"
-        pacman-key --lsign-key "${_archzfs_key}"
+        if ! (
+            pacman-key --recv-keys "${_archzfs_key}" &&
+                pacman-key --finger "${_archzfs_key}" &&
+                pacman-key --lsign-key "${_archzfs_key}"
+        ); then
+            printf "blackarch> failed, exiting"
+            exit 3
+        fi
 
         __separator
         __confirm "zfs"
